+++ package.json
@@ -3,10 +3,11 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
-    "start": "node server.js",
+    "start": "node dist/server.js",
     "cy:open": "cypress open",
     "cy:run": "cypress run",
-    "test": "npm run cy:run"
+    "test": "npm run cy:run",
+    "build": "tsc"
   },
   "keywords": [],
   "author": "",
@@ -15,7 +16,11 @@
   "devDependencies": {
     "@badeball/cypress-cucumber-preprocessor": "^22.0.1",
     "@bahmutov/cypress-esbuild-preprocessor": "^2.2.5",
+    "@types/cypress": "^0.1.6",
+    "@types/node": "^22.15.30",
     "cypress": "^14.4.1",
-    "cypress-cucumber-preprocessor": "^4.3.1"
+    "cypress-cucumber-preprocessor": "^4.3.1",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.8.3"
   }
 }
